requirements.txt
streamlit
torch
torchvision
ultralytics
opencv-python
numpy
Pillow
streamlit
opencv-python
numpy
ultralytics
pillow
pandas
matplotlib
plotly
folium
streamlit-folium
streamlit-option-menu
